fix color for three
