+++ packages/core/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './math'
 export * from './object'
 export * from './string'
+export * from './color'
