+++ packages/core/src/utils/string.ts
@@ -4,3 +4,15 @@ export function parseUnitValue(value: number | string): [number] | [number, stri
   const unit: string = value.substring(('' + _value).length)
   return [_value, unit]
 }
+
+export function substringMatch(str: string, from: string, to?: string) {
+  const pos = str.indexOf(from)
+  if (pos !== -1) {
+    if (to) {
+      return str.substring(pos + from.length, str.indexOf(to))
+    }
+    return str.substring(0, pos)
+  }
+
+  return ''
+}
