+++ packages/three/src/adapters/euler.ts
@@ -1,12 +1,13 @@
-import { Adapter } from '@animini/core'
 import { Euler, Vector3 } from 'three'
+import { ThreeAdapter } from '../types'
 
-export const euler: Adapter = {
+export const euler: ThreeAdapter = {
   parseInitial(euler: Euler) {
     const v = new Vector3()
     return v.setFromEuler(euler) as any as Record<string, number>
   },
-  onChange(element, key, value) {
-    element[key].setFromVector3(value)
+  onChange(value, key, element) {
+    // @ts-ignore
+    element![key].setFromVector3(value)
   }
 }
